hamravesh/requirements.txt
django
django-rest-framework
celery
pyamqp